Refactoring backend files
