+++ backend/src/main/java/com/training/feedbacktool/dto/SurveyResponse.java
@@ -1,4 +1,4 @@
-package com.training.feedbacktool.survey.api.dto;
+package com.training.feedbacktool.dto;
 
 import java.time.Instant;
